LayoutTests/imported/w3c/web-platform-tests/domxpath/node-sets-expected.txt


FAIL | operator should evaluate both sides of expressions with the same context node assert_equals: expected 2 but got 1
